requirements.txt
flask
pymysql
python-dotenv
kafka-python==2.0.2
cryptography
lz4
six